Test the library against PyPy
This fixes several issues when installed in PyPy:

- Some macros were undefined.
- Some code relied on CPython internals and/or refcounting
- Some C functions are buggy in PyPy
- A few things work differently in PyPy

It also adds a workflow to test against PyPy3.5 7.0.0 and
PyPy3.6 7.2.0. Unfortunately PyPy doesn't find the C compiler
on Windows so windows is currently untested. Also something
(maybe the test runner or a dependency) runs into a segfault
or other exception during the execution on Linux so on Linux
the library is only installed not tested.