Github-Actions: Use latest versions of OS where applicable
With the deprecation of macos 10.14 it is necessary to swap
mac to latest. The other changes to latest OS are just convenience.

Also move some of the tests to the faster (and probably more availabl)
ubuntu OS.

Also remove manual installation of Python 3.8 because it's now included
in the default software.