Fixed a fatal error in grouper.
The tp_next method of grouper had a reference counting error
which could lead to segmentation faults when the instance was
deallocated. Unfortunately this wasn't catched by the tests
because the memory location wasn't re-used in the meantime.

So this commit adds an assertion in the tp_dealloc handler
which is now tested on Travis and GitHub Actions and the
reference counting bug has been fixed.

The installation instruction documentation was updated to
include how this can be done locally.